WIP: Moved components from shared to components. Added jsconfig tto handle updating imports when moving
