+++ src/components/SVGLoad/SVG.js
@@ -1,11 +1,11 @@
 import React from 'react';
 
-import { useStyles } from './styles/SVGStyle'
+import { useStyles } from './SVGStyle'
 
 export default function SVGLoad(props){
     const PicSVG = props.image;
     const classes = useStyles();
-    
+
     return(
         <PicSVG className={classes.svgSettings} />
     )
