+++ src/components/SVGLoad/SVGStyle.js
